implement cv2 reply methods and move message based replies to own class
