+++ core/src/main/java/io/github/kaktushose/jdac/dispatching/reply/ConfigurableReply.java
@@ -2,48 +2,27 @@ package io.github.kaktushose.jdac.dispatching.reply;
 
 import io.github.kaktushose.jdac.JDACBuilder;
 import io.github.kaktushose.jdac.annotations.interactions.ReplyConfig;
-import io.github.kaktushose.jdac.definitions.interactions.CustomId;
 import io.github.kaktushose.jdac.definitions.interactions.InteractionDefinition;
-import io.github.kaktushose.jdac.definitions.interactions.InteractionRegistry;
-import io.github.kaktushose.jdac.definitions.interactions.ModalDefinition;
-import io.github.kaktushose.jdac.definitions.interactions.component.ButtonDefinition;
-import io.github.kaktushose.jdac.definitions.interactions.component.ComponentDefinition;
-import io.github.kaktushose.jdac.definitions.interactions.component.menu.SelectMenuDefinition;
-import io.github.kaktushose.jdac.dispatching.reply.dynamic.ButtonComponent;
-import io.github.kaktushose.jdac.dispatching.reply.dynamic.internal.UnspecificComponent;
-import io.github.kaktushose.jdac.dispatching.reply.dynamic.menu.EntitySelectMenuComponent;
-import io.github.kaktushose.jdac.dispatching.reply.dynamic.menu.StringSelectComponent;
+import io.github.kaktushose.jdac.dispatching.reply.internal.ComponentReplyAction;
 import io.github.kaktushose.jdac.dispatching.reply.internal.MessageReplyAction;
 import io.github.kaktushose.jdac.embeds.Embed;
 import io.github.kaktushose.jdac.embeds.EmbedConfig;
-import io.github.kaktushose.jdac.exceptions.InternalException;
-import io.github.kaktushose.jdac.exceptions.internal.JDACException;
 import io.github.kaktushose.jdac.message.i18n.I18n;
 import io.github.kaktushose.jdac.message.placeholder.Entry;
-import net.dv8tion.jda.api.components.ActionComponent;
 import net.dv8tion.jda.api.components.MessageTopLevelComponent;
 import net.dv8tion.jda.api.components.actionrow.ActionRow;
-import net.dv8tion.jda.api.components.actionrow.ActionRowChildComponent;
-import net.dv8tion.jda.api.components.buttons.Button;
-import net.dv8tion.jda.api.components.selections.EntitySelectMenu;
-import net.dv8tion.jda.api.components.selections.SelectOption;
-import net.dv8tion.jda.api.components.selections.StringSelectMenu;
-import net.dv8tion.jda.api.components.selections.StringSelectMenu.Builder;
 import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.requests.RestAction;
 import net.dv8tion.jda.api.utils.messages.MessageCreateBuilder;
-import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.Locale;
 import java.util.function.Consumer;
-import java.util.function.Function;
 
 import static io.github.kaktushose.jdac.dispatching.context.internal.RichInvocationContext.*;
-import static io.github.kaktushose.jdac.message.placeholder.Entry.entry;
 
 /// Builder for sending messages based on a [GenericInteractionCreateEvent] that supports adding components to
 /// messages and changing the [InteractionDefinition.ReplyConfig].
@@ -64,27 +43,22 @@ import static io.github.kaktushose.jdac.message.placeholder.Entry.entry;
 ///     }
 /// }
 /// ```
-public sealed class ConfigurableReply permits SendableReply {
+public final class ConfigurableReply {
 
     private static final Logger log = LoggerFactory.getLogger(ConfigurableReply.class);
-    protected final MessageReplyAction replyAction;
+    private boolean ephemeral;
+    private boolean editReply;
+    private boolean keepComponents;
+    private boolean keepSelections;
 
     /// Constructs a new ConfigurableReply.
     ///
-    /// @param replyAction the underlying [MessageReplyAction]
-    public ConfigurableReply(MessageReplyAction replyAction) {
-        this.replyAction = replyAction;
-    }
-
-    public CV2Reply cv2(MessageTopLevelComponent component, MessageTopLevelComponent... components) {
-        return new CV2Reply(component, components, getReplyConfig());
-    }
-
-    /// Constructs a new ConfigurableReply.
-    ///
-    /// @param configurableReply the [ConfigurableReply] to copy
-    public ConfigurableReply(ConfigurableReply configurableReply) {
-        this.replyAction = configurableReply.replyAction;
+    /// @param replyConfig the underlying [InteractionDefinition.ReplyConfig]
+    public ConfigurableReply(InteractionDefinition.ReplyConfig replyConfig) {
+        ephemeral = replyConfig.ephemeral();
+        editReply = replyConfig.editReply();
+        keepComponents = replyConfig.keepComponents();
+        keepSelections = replyConfig.keepSelections();
     }
 
     /// Whether to send ephemeral replies. Default value is `false`.
@@ -100,7 +74,7 @@ public sealed class ConfigurableReply permits SendableReply {
     /// @param ephemeral `true` if to send ephemeral replies
     /// @return the current instance for fluent interface
     public ConfigurableReply ephemeral(boolean ephemeral) {
-        replyAction.ephemeral(ephemeral);
+        this.ephemeral = ephemeral;
         return this;
     }
 
@@ -114,7 +88,7 @@ public sealed class ConfigurableReply permits SendableReply {
     /// @param editReply `true` if to keep the original components
     /// @return the current instance for fluent interface
     public ConfigurableReply editReply(boolean editReply) {
-        replyAction.editReply(editReply);
+        this.editReply = editReply;
         return this;
     }
 
@@ -130,28 +104,14 @@ public sealed class ConfigurableReply permits SendableReply {
     /// @param keepComponents `true` if to edit the original method
     /// @return the current instance for fluent interface
     public ConfigurableReply keepComponents(boolean keepComponents) {
-        replyAction.keepComponents(keepComponents);
+        this.keepComponents = keepComponents;
         return this;
     }
 
     /// Whether to keep the selections of a string select menu when sending edits. This setting only has an effect with
     /// [#keepComponents(boolean)] `true`.
     public ConfigurableReply keepSelections(boolean keepSelections) {
-        replyAction.keepSelections(keepSelections);
-        return this;
-    }
-
-    /// Access the underlying [MessageCreateBuilder] for configuration steps not covered by [ConfigurableReply].
-    ///
-    /// This method exposes the internal [MessageCreateBuilder] used by JDA-Commands. Modifying fields that
-    /// are also manipulated by the Reply API, like content or embeds, may lead to unexpected behaviour.
-    ///
-    /// ## Example:
-    /// ```
-    /// event.with().builder(builder -> builder.setFiles(myFile)).reply("Hello World!");
-    ///```
-    public ConfigurableReply builder(Consumer<MessageCreateBuilder> builder) {
-        replyAction.builder(builder);
+        this.keepSelections = keepSelections;
         return this;
     }
 
@@ -165,7 +125,24 @@ public sealed class ConfigurableReply permits SendableReply {
     ///
     /// This might throw [RuntimeException]s if JDA fails to send the message.
     public Message reply(String message, Entry... placeholder) {
-        return replyAction.reply(message, placeholder);
+        return new MessageReplyAction(replyConfig()).reply(message, placeholder);
+    }
+
+    public Message reply(MessageTopLevelComponent component, MessageTopLevelComponent... components) {
+        return new ComponentReplyAction(getReplyConfig(), component, components).reply();
+    }
+
+    /// Access the underlying [MessageCreateBuilder] for configuration steps not covered by [ConfigurableReply].
+    ///
+    /// This method exposes the internal [MessageCreateBuilder] used by JDA-Commands. Modifying fields that
+    /// are also manipulated by the Reply API, like content or embeds, may lead to unexpected behaviour.
+    ///
+    /// ## Example:
+    /// ```
+    /// event.with().builder(builder -> builder.setFiles(myFile)).reply("Hello World!");
+    ///```
+    public MessageReply builder(Consumer<MessageCreateBuilder> consumer) {
+        return newReply().builder(consumer);
     }
 
     /// Acknowledgement of this event with one or more [Embed]s.
@@ -174,8 +151,10 @@ public sealed class ConfigurableReply permits SendableReply {
     ///
     /// @param embeds the name of the [Embed]s to send
     /// @return a new [SendableReply]
-    public SendableReply embeds(String... embeds) {
-        return embeds(Arrays.stream(embeds).map(it -> getFramework().embeds().get(it, getJdaEvent().getUserLocale().toLocale())).toArray(Embed[]::new));
+    public MessageReply embeds(String... embeds) {
+        return embeds(Arrays.stream(embeds)
+                .map(it -> getFramework().embeds().get(it, getJdaEvent().getUserLocale().toLocale()))
+                .toArray(Embed[]::new));
     }
 
     /// Acknowledgement of this event with one or more [Embed]s.
@@ -184,9 +163,8 @@ public sealed class ConfigurableReply permits SendableReply {
     ///
     /// @param embeds the [Embed]s to send
     /// @return a new [SendableReply]
-    public SendableReply embeds(Embed... embeds) {
-        replyAction.addEmbeds(Arrays.stream(embeds).map(Embed::build).toArray(MessageEmbed[]::new));
-        return new SendableReply(this);
+    public MessageReply embeds(Embed... embeds) {
+        return newReply().embeds(embeds);
     }
 
     /// Acknowledgement of this event with an [Embed].
@@ -196,11 +174,8 @@ public sealed class ConfigurableReply permits SendableReply {
     /// @param embed    the name of the [Embed] to send
     /// @param consumer a [Consumer] allowing direct modification of the [Embed] before sending it.
     /// @return a new [SendableReply]
-    public SendableReply embeds(String embed, Consumer<Embed> consumer) {
-        Embed resolved = getFramework().embeds().get(embed, getJdaEvent().getUserLocale().toLocale());
-        consumer.accept(resolved);
-        replyAction.addEmbeds(resolved.build());
-        return new SendableReply(this);
+    public MessageReply embeds(String embed, Consumer<Embed> consumer) {
+        return newReply().embeds(embed, consumer);
     }
 
     /// Acknowledgement of this event with an [Embed].
@@ -211,11 +186,8 @@ public sealed class ConfigurableReply permits SendableReply {
     /// @param entry   the placeholders to use. See [Embed#placeholders(Entry...)]
     /// @param entries the placeholders to use. See [Embed#placeholders(Entry...)]
     /// @return a new [SendableReply]
-    public SendableReply embeds(String embed, Entry entry, Entry... entries) {
-        Embed resolved = getFramework().embeds().get(embed, getJdaEvent().getUserLocale().toLocale());
-        resolved.placeholders(entry).placeholders(entries);
-        replyAction.addEmbeds(resolved.build());
-        return new SendableReply(this);
+    public MessageReply embeds(String embed, Entry entry, Entry... entries) {
+        return newReply().embeds(embed, entry, entries);
     }
 
     /// Adds an [ActionRow] to the reply and adds the passed components to it.
@@ -241,14 +213,11 @@ public sealed class ConfigurableReply permits SendableReply {
     ///     }
     ///  }
     ///```
-    /// @param components the name of the components to add
-    /// @return the current instance for fluent interface
-    public SendableReply components(String... components) {
+    public MessageReply components(String... components) {
         return components(Arrays.stream(components).map(Component::enabled).toArray(Component[]::new));
     }
 
     /// Adds an [ActionRow] to the reply and adds the passed [Component] to it.
-    ///
     ///
     /// ### Example:
     /// ```
@@ -267,118 +236,17 @@ public sealed class ConfigurableReply permits SendableReply {
     ///  }
     ///```
     /// @see Component
-    /// @param components the [Component] to add
-    /// @return the current instance for fluent interface
-    public SendableReply components(Component<?, ?, ?, ?>... components) {
-        List<ActionRowChildComponent> items = new ArrayList<>();
-        for (Component<?, ?, ?, ?> component : components) {
-            var className = component.origin().map(Class::getName)
-                    .orElseGet(() -> getInvocationContext().definition().methodDescription().declaringClass().getName());
-            String definitionId = InteractionDefinition.createDefinitionId(className, component.name());
-
-            if (replyAction.components()
-                    .stream()
-                    .map(ActionComponent::getCustomId)
-                    .filter(Objects::nonNull)
-                    .map(CustomId::fromMerged)
-                    .anyMatch(customId -> customId.definitionId().equals(definitionId))) {
-                throw new IllegalArgumentException(
-                        JDACException.errorMessage("duplicate-component", entry("method", "%s#%s".formatted(className, component.name())))
-                );
-            }
-
-            var definition = findDefinition(component, definitionId, className);
-
-            ActionRowChildComponent item = switch (definition) {
-                case ButtonDefinition buttonDefinition -> {
-                    var button = buttonDefinition.toJDAEntity().withDisabled(!component.enabled());
-                    //only assign ids to non-link buttons
-                    yield button.getUrl() == null ? button.withCustomId(createId(definition, component.independent()).merged()) : button;
-                }
-
-                case SelectMenuDefinition<?> menuDefinition -> {
-                    var menu = menuDefinition.toJDAEntity(createId(definition, component.independent()));
-                    yield menu.withDisabled(!component.enabled());
-                }
-            };
-
-            item = switch (component) {
-                case ButtonComponent buttonComponent -> buttonComponent.callback().apply((Button) item);
-                case EntitySelectMenuComponent entitySelectMenuComponent ->
-                        entitySelectMenuComponent.callback().apply(((EntitySelectMenu) item).createCopy()).build();
-                case StringSelectComponent stringSelectComponent ->
-                        stringSelectComponent.callback().apply(((StringSelectMenu) item).createCopy()).build();
-                case UnspecificComponent unspecificComponent -> unspecificComponent.callback().apply(item);
-            };
-
-            item = resolve(item, component);
-
-            items.add(item);
-
-            log.debug("Reply Debug: Adding component \"{}\" to the reply", definition.displayName());
-        }
-
-        if (!items.isEmpty()) {
-            replyAction.addComponents(ActionRow.of(items));
-        }
-        return new SendableReply(this);
+    public MessageReply components(Component<?, ?, ?, ?>... components) {
+        return newReply().components(components);
     }
 
-    private ActionRowChildComponent resolve(ActionRowChildComponent item, Component<?, ?, ?, ?> component) {
-        return switch (item) {
-            case Button button -> button.withLabel(resolve(button.getLabel(), component));
-            case EntitySelectMenu menu -> menu.createCopy().setPlaceholder(orNull(menu.getPlaceholder(),p -> resolve(p, component))).build();
-            case StringSelectMenu menu -> {
-                Builder copy = menu.createCopy();
-                List<SelectOption> localized = copy.getOptions()
-                        .stream()
-                        .map(option -> option.withDescription(orNull(option.getDescription(), d -> resolve(d, component)))
-                                .withLabel(resolve(option.getLabel(), component)))
-                        .toList();
-                copy.getOptions().clear();
-                copy.addOptions(localized);
-                copy.setPlaceholder(orNull(copy.getPlaceholder(), p -> resolve(p, component)));
-                yield copy.build();
-            }
-            default -> throw new InternalException("default-switch");
-        };
+    private InteractionDefinition.ReplyConfig replyConfig() {
+        log.debug("Reply Debug: [Runtime={}]", getRuntime().id());
+        return new InteractionDefinition.ReplyConfig(ephemeral, editReply, keepComponents, keepSelections);
     }
 
-    @Nullable
-    private <T> T orNull(@Nullable T val, Function<T, T> func) {
-        if (val == null) return null;
-        return func.apply(val);
-    }
-
-    private String resolve(String key, Component<?, ?, ?, ?> component) {
-        return getFramework().messageResolver().resolve(key, getJdaEvent().getUserLocale().toLocale(), component.placeholder());
-    }
-
-    private <D extends ComponentDefinition<?>, T extends Component<T, ?, ?, D>> D findDefinition(Component<T, ?, ?, D> component, String definitionId, String className) {
-        InteractionRegistry registry = getFramework().interactionRegistry();
-
-        try {
-            // this cast is effective safe
-            D definition = registry.find(component.definitionClass(), false, it ->
-                    it.definitionId().equals(definitionId)
-            );
-
-            return component.build(definition);
-        } catch (IllegalArgumentException e) { // only check if search failed
-            Collection<ModalDefinition> found = registry.find(ModalDefinition.class, it -> it.definitionId().equals(definitionId));
-            if (!found.isEmpty()) {
-                throw new IllegalArgumentException(
-                        JDACException.errorMessage("modal-as-component", entry("method", "%s#%s".formatted(className, component.name())))
-                );
-            }
-            throw e;
-        }
-    }
-
-    private CustomId createId(InteractionDefinition definition, boolean staticComponent) {
-        return staticComponent
-                ? CustomId.independent(definition.definitionId())
-                : new CustomId(getRuntime().id(), definition.definitionId());
+    private MessageReply newReply() {
+        return new MessageReply(replyConfig());
     }
 }
